+++ postal_address/__init__.py
@@ -17,4 +17,8 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
+
 __version__ = '0.3.0.dev'
+
+
+from .address import Address